+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 - `ext:install` now supports `--force` and `--non-interactive` flags.
 - Fixes a crash when customers deploy an empty Cloud Functions project (#3705)
 - Fixes (and implements) `--no-authorized-domains`, skipping syncing with Firebase Auth, when deploying to a Firebase Hosting channel (#3740).
+- Uses public v1 AppDistribution API.
+- Adds `appdistribution:testers:add` and `appdistribution:testers:remove` commands (#3072).